BUGFIX: use the env we create when evaluating a block
This happened when I switched from the jlox style of storing the current
environment as a variable and setting it as appropriate when entering or
exiting a scope to the Lisp / Scheme style of passing the environment to
eval (see the sidenote chapeter 8.5.2).  I renamed the block environment
(previously called "env"), but missed changing it in the next line where
it was actually used (hey, it was late).